split hid demod/awid demod to respective files
NOTE: see changelog.md for cli changes
